Allow for an ignored local settings file.
Even though settings.json is in the .gitignore file git does not ignore
modifications to this file. This change makes expressCart detect the
existence of a file called settings-local.json. When this file exists,
it will be used for reading and writing settings instead of the usual
settings.json. This will help to prevent accidental leakage of user
secrets.